+++ packages/plasma-asdk/src/components/Radiobox/Radiobox.config.ts
@@ -13,7 +13,7 @@ export const config = {
         size: {
             s: css`
                 ${radioboxTokens.margin}: 0;
-                ${radioboxTokens.triggerMargin}: 0;
+                ${radioboxTokens.triggerPadding}: 0;
                 ${radioboxTokens.triggerSize}: 1rem;
                 ${radioboxTokens.triggerBorderRadius}: 1.125rem;
                 ${radioboxTokens.ellipseSize}: 0.5rem;
@@ -35,7 +35,7 @@ export const config = {
             `,
             m: css`
                 ${radioboxTokens.margin}: 0;
-                ${radioboxTokens.triggerMargin}: 0.063rem 0;
+                ${radioboxTokens.triggerPadding}: 0.063rem;
                 ${radioboxTokens.triggerSize}: 1.375rem;
                 ${radioboxTokens.triggerBorderRadius}: 1.125rem;
                 ${radioboxTokens.ellipseSize}: 0.625rem;
