+++ packages/plasma-b2c/src/components/Checkbox/Checkbox.config.ts
@@ -7,17 +7,15 @@ export const config = {
         focused: 'true',
     },
     variations: {
-        /**
-         * Размер контрола.
-         */
         size: {
             s: css`
                 ${checkboxTokens.margin}: 0;
-                ${checkboxTokens.triggerMargin}: 0.063rem 0;
+                ${checkboxTokens.triggerPadding}: 0.0625rem;
                 ${checkboxTokens.triggerSize}: 0.875rem;
                 ${checkboxTokens.triggerBorderRadius}: 0.25rem;
                 ${checkboxTokens.contentTopOffset}: 0;
                 ${checkboxTokens.contentLeftOffset}: 0.5rem;
+                ${checkboxTokens.descriptionMarginTop}: 0.125rem;
                 ${checkboxTokens.labelFontFamily}: var(--plasma-typo-body-s-font-family);
                 ${checkboxTokens.labelFontSize}: var(--plasma-typo-body-s-font-size);
                 ${checkboxTokens.labelFontStyle}: var(--plasma-typo-body-s-font-style);
@@ -25,7 +23,6 @@ export const config = {
                 ${checkboxTokens.labelLetterSpacing}: var(--plasma-typo-body-s-letter-spacing);
                 ${checkboxTokens.labelLineHeight}: var(--plasma-typo-body-s-line-height);
                 ${checkboxTokens.descriptionFontFamily}: var(--plasma-typo-body-xs-font-family);
-                ${checkboxTokens.descriptionMarginTop}: 0.125rem;
                 ${checkboxTokens.descriptionFontSize}: var(--plasma-typo-body-xs-font-size);
                 ${checkboxTokens.descriptionFontStyle}: var(--plasma-typo-body-xs-font-style);
                 ${checkboxTokens.descriptionFontWeight}: var(--plasma-typo-body-xs-font-weight);
@@ -34,11 +31,12 @@ export const config = {
             `,
             m: css`
                 ${checkboxTokens.margin}: 0;
-                ${checkboxTokens.triggerMargin}: 0.125rem 0;
+                ${checkboxTokens.triggerPadding}: 0.125rem;
                 ${checkboxTokens.triggerSize}: 1.25rem;
                 ${checkboxTokens.triggerBorderRadius}: 0.375rem;
                 ${checkboxTokens.contentTopOffset}: 0.125rem;
                 ${checkboxTokens.contentLeftOffset}: 0.75rem;
+                ${checkboxTokens.descriptionMarginTop}: 0.125rem;
                 ${checkboxTokens.labelFontFamily}: var(--plasma-typo-body-m-font-family);
                 ${checkboxTokens.labelFontSize}: var(--plasma-typo-body-m-font-size);
                 ${checkboxTokens.labelFontStyle}: var(--plasma-typo-body-m-font-style);
@@ -46,7 +44,6 @@ export const config = {
                 ${checkboxTokens.labelLetterSpacing}: var(--plasma-typo-body-m-letter-spacing);
                 ${checkboxTokens.labelLineHeight}: var(--plasma-typo-body-m-line-height);
                 ${checkboxTokens.descriptionFontFamily}: var(--plasma-typo-body-s-font-family);
-                ${checkboxTokens.descriptionMarginTop}: 0.125rem;
                 ${checkboxTokens.descriptionFontSize}: var(--plasma-typo-body-s-font-size);
                 ${checkboxTokens.descriptionFontStyle}: var(--plasma-typo-body-s-font-style);
                 ${checkboxTokens.descriptionFontWeight}: var(--plasma-typo-body-s-font-weight);
@@ -54,9 +51,6 @@ export const config = {
                 ${checkboxTokens.descriptionLineHeight}: var(--plasma-typo-body-s-line-height);
             `,
         },
-        /**
-         * Вид контрола.
-         */
         view: {
             accent: css`
                 ${checkboxTokens.fillColor}: var(--text-accent);
