+++ packages/plasma-b2c/src/components/Radiobox/Radiobox.config.ts
@@ -13,12 +13,13 @@ export const config = {
         size: {
             s: css`
                 ${radioboxTokens.margin}: 0;
-                ${radioboxTokens.triggerMargin}: 0.188rem 0;
+                ${radioboxTokens.triggerPadding}: 0.0625rem 0;
                 ${radioboxTokens.triggerSize}: 1rem;
                 ${radioboxTokens.triggerBorderRadius}: 1.125rem;
                 ${radioboxTokens.ellipseSize}: 0.5rem;
-                ${radioboxTokens.contentTopOffset}: 0.1rem;
+                ${radioboxTokens.contentTopOffset}: 0;
                 ${radioboxTokens.contentLeftOffset}: 0.5rem;
+                ${radioboxTokens.descriptionMarginTop}: 0.125rem;
                 ${radioboxTokens.labelFontFamily}: var(--plasma-typo-body-s-font-family);
                 ${radioboxTokens.labelFontSize}: var(--plasma-typo-body-s-font-size);
                 ${radioboxTokens.labelFontStyle}: var(--plasma-typo-body-s-font-style);
@@ -26,7 +27,6 @@ export const config = {
                 ${radioboxTokens.labelLetterSpacing}: var(--plasma-typo-body-s-letter-spacing);
                 ${radioboxTokens.labelLineHeight}: var(--plasma-typo-body-s-line-height);
                 ${radioboxTokens.descriptionFontFamily}: var(--plasma-typo-body-xs-font-family);
-                ${radioboxTokens.descriptionMarginTop}: 0.125rem;
                 ${radioboxTokens.descriptionFontSize}: var(--plasma-typo-body-xs-font-size);
                 ${radioboxTokens.descriptionFontStyle}: var(--plasma-typo-body-xs-font-style);
                 ${radioboxTokens.descriptionFontWeight}: var(--plasma-typo-body-xs-font-weight);
@@ -35,12 +35,13 @@ export const config = {
             `,
             m: css`
                 ${radioboxTokens.margin}: 0;
-                ${radioboxTokens.triggerMargin}: 0.063rem 0;
+                ${radioboxTokens.triggerPadding}: 0.0625rem;
                 ${radioboxTokens.triggerSize}: 1.375rem;
                 ${radioboxTokens.triggerBorderRadius}: 1.125rem;
                 ${radioboxTokens.ellipseSize}: 0.625rem;
                 ${radioboxTokens.contentTopOffset}: 0.125rem;
                 ${radioboxTokens.contentLeftOffset}: 0.625rem;
+                ${radioboxTokens.descriptionMarginTop}: 0.125rem;
                 ${radioboxTokens.labelFontFamily}: var(--plasma-typo-body-m-font-family);
                 ${radioboxTokens.labelFontSize}: var(--plasma-typo-body-m-font-size);
                 ${radioboxTokens.labelFontStyle}: var(--plasma-typo-body-m-font-style);
@@ -48,7 +49,6 @@ export const config = {
                 ${radioboxTokens.labelLetterSpacing}: var(--plasma-typo-body-m-letter-spacing);
                 ${radioboxTokens.labelLineHeight}: var(--plasma-typo-body-m-line-height);
                 ${radioboxTokens.descriptionFontFamily}: var(--plasma-typo-body-s-font-family);
-                ${radioboxTokens.descriptionMarginTop}: 0.125rem;
                 ${radioboxTokens.descriptionFontSize}: var(--plasma-typo-body-s-font-size);
                 ${radioboxTokens.descriptionFontStyle}: var(--plasma-typo-body-s-font-style);
                 ${radioboxTokens.descriptionFontWeight}: var(--plasma-typo-body-s-font-weight);
