+++ packages/plasma-new-hope/src/components/Radiobox/Radiobox.styles.ts
@@ -31,7 +31,7 @@ const mappingOverride = `
     ${checkboxTokens.contentLeftOffset}: var(${tokens.contentLeftOffset});
     ${checkboxTokens.triggerSize}: var(${tokens.triggerSize});
     ${checkboxTokens.triggerBorderRadius}: var(${tokens.triggerBorderRadius});
-    ${checkboxTokens.triggerMargin}: var(${tokens.triggerMargin});
+    ${checkboxTokens.triggerPadding}: var(${tokens.triggerPadding});
 `;
 
 export const StyledEllipse = styled.div`
