+++ packages/plasma-new-hope/src/components/Radiobox/Radiobox.tsx
@@ -10,6 +10,7 @@ import {
     StyledLabel,
     StyledContent,
     StyledTrigger,
+    StyledTriggerWrapper,
 } from '../Checkbox/Checkbox.styles';
 import { BaseboxProps } from '../Checkbox/Checkbox.types';
 import { cx } from '../../utils';
@@ -73,9 +74,11 @@ export const radioboxRoot = (Root: RootProps<HTMLInputElement, RadioboxProps>) =
                     tabIndex={canFocused}
                 />
                 <StyledContentWrapper htmlFor={radioboxId} className={singleLineClass}>
-                    <StyledTrigger className={classes.radioboxTrigger}>
-                        <StyledEllipse />
-                    </StyledTrigger>
+                    <StyledTriggerWrapper>
+                        <StyledTrigger className={classes.radioboxTrigger}>
+                            <StyledEllipse />
+                        </StyledTrigger>
+                    </StyledTriggerWrapper>
                     {hasContent && (
                         <StyledContent className={singleLineClass}>
                             {label && (
