+++ packages/plasma-web/src/components/Checkbox/Checkbox.config.ts
@@ -10,7 +10,7 @@ export const config = {
         size: {
             s: css`
                 ${checkboxTokens.margin}: 0;
-                ${checkboxTokens.triggerMargin}: 0.063rem 0;
+                ${checkboxTokens.triggerPadding}: 0.0625rem;
                 ${checkboxTokens.triggerSize}: 0.875rem;
                 ${checkboxTokens.triggerBorderRadius}: 0.25rem;
                 ${checkboxTokens.contentTopOffset}: 0;
@@ -31,7 +31,7 @@ export const config = {
             `,
             m: css`
                 ${checkboxTokens.margin}: 0;
-                ${checkboxTokens.triggerMargin}: 0.125rem 0;
+                ${checkboxTokens.triggerPadding}: 0.125rem;
                 ${checkboxTokens.triggerSize}: 1.25rem;
                 ${checkboxTokens.triggerBorderRadius}: 0.375rem;
                 ${checkboxTokens.contentTopOffset}: 0.125rem;
@@ -61,6 +61,14 @@ export const config = {
             `,
             // deprecated
             default: css`
+                ${checkboxTokens.fillColor}: var(--text-primary);
+                ${checkboxTokens.iconColor}: var(--inverse-text-primary);
+                ${checkboxTokens.descriptionColor}: var(--text-secondary);
+                ${checkboxTokens.triggerBackgroundColor}: transparent;
+                ${checkboxTokens.triggerBorderColor}: var(--text-secondary);
+            `,
+            // deprecated
+            primary: css`
                 ${checkboxTokens.fillColor}: var(--text-primary);
                 ${checkboxTokens.iconColor}: var(--inverse-text-primary);
                 ${checkboxTokens.descriptionColor}: var(--text-secondary);
